Update Antonio_Ramos_Documentation
Let me know what you think about the following plan. 
All the data is now projecting correctly and I am going to run the K-function Local_3Stage.py program to see if we get a good cluster file after 999 simulations. That will take some time so while that is happening I am also going to clean up all the data we have on the VM and G drive. I will delete old data that isn't being used and only keep the stuff that is used by the K-function program or one of the scripts that prepares data for the K-function program. Moreover, I will programmatize the creation of the CA_Block_New.shp data.